# Reduce cuda copying