--- a/src/icrar/leap-accelerate/algorithm/cpu/PhaseRotate.cc
+++ b/src/icrar/leap-accelerate/algorithm/cpu/PhaseRotate.cc
@@ -126,7 +126,7 @@ namespace cpu
             BOOST_LOG_TRIVIAL(info) << "Processing direction " << i;
             metadata.SetDD(directions[i]);
             metadata.CalcUVW();
-            metadata.avg_data.setConstant(std::complex<double>(0.0,0.0));
+            metadata.GetAvgData().setConstant(std::complex<double>(0.0,0.0));
             icrar::cpu::PhaseRotate(metadata, directions[i], input_queues[i], output_integrations[i], output_calibrations[i]);
         }
 
@@ -151,8 +151,9 @@ namespace cpu
             icrar::cpu::RotateVisibilities(integration, metadata);
             output_integrations.emplace_back(direction, integration.GetIntegrationNumber(), boost::none);
         }
-        BOOST_LOG_TRIVIAL(trace) << "avg_data: " << pretty_matrix(metadata.avg_data);
+
 #ifdef TRACE
+        BOOST_LOG_TRIVIAL(trace) << "avg_data: " << pretty_matrix(metadata.GetAvgData());
         {
             std::ofstream file;
             file.open("avg_data.txt");
@@ -161,7 +162,7 @@ namespace cpu
         }
 #endif
 
-        auto avg_data_angles = metadata.avg_data.unaryExpr([](std::complex<double> c) -> Radians { return std::arg(c); });
+        auto avg_data_angles = metadata.GetAvgData().unaryExpr([](std::complex<double> c) -> Radians { return std::arg(c); });
 
         // TODO: reference antenna should be included and set to 0?
         auto cal_avg_data = icrar::cpu::VectorRangeSelect(avg_data_angles, metadata.GetI1(), 0); // 1st pol only
@@ -169,7 +170,7 @@ namespace cpu
         // cal_avg_data(cal_avg_data.size() - 1) = 0.0;
         Eigen::VectorXd cal1 = metadata.GetAd1() * cal_avg_data;
 
-        Eigen::MatrixXd dInt = Eigen::MatrixXd::Zero(metadata.GetI().size(), metadata.avg_data.cols());
+        Eigen::MatrixXd dInt = Eigen::MatrixXd::Zero(metadata.GetI().size(), metadata.GetAvgData().cols());
         Eigen::MatrixXd avg_data_slice = icrar::cpu::MatrixRangeSelect(avg_data_angles, metadata.GetI(), Eigen::all);
         for(int n = 0; n < metadata.GetI().size(); ++n)
         {
@@ -193,7 +194,7 @@ namespace cpu
 
         metadata.CalcUVW();
 
-        const auto polar_direction = icrar::ToPolar(metadata.direction);
+        const auto polar_direction = icrar::ToPolar(metadata.GetDirection());
         
         // loop over smeared baselines
         for(size_t baseline = 0; baseline < integration.GetBaselines(); ++baseline)
@@ -236,7 +237,7 @@ namespace cpu
                 {
                     for(int polarization = 0; polarization < metadata.GetConstants().num_pols; ++polarization)
                     {
-                        metadata.avg_data(md_baseline, polarization) += integration_data(polarization, baseline, channel);
+                        metadata.GetAvgData()(md_baseline, polarization) += integration_data(polarization, baseline, channel);
                     }
                 }
             }