--- a/src/icrar/leap-accelerate/algorithm/cpu/PhaseRotate.cc
+++ b/src/icrar/leap-accelerate/algorithm/cpu/PhaseRotate.cc
@@ -157,7 +157,7 @@ namespace cpu
         {
             std::ofstream file;
             file.open("avg_data.txt");
-            file << metadata.avg_data << std::endl;
+            file << metadata.GetAvgData() << std::endl;
             file.close();
         }
 #endif