--- a/src/icrar/leap-accelerate/model/cpu/MetaData.cc
+++ b/src/icrar/leap-accelerate/model/cpu/MetaData.cc
@@ -61,7 +61,7 @@ namespace cpu
 
         if(metadata.dd.is_initialized())
         {
-            dd = ToMatrix<double, 3, 3>(metadata.dd.value());
+            m_dd = ToMatrix<double, 3, 3>(metadata.dd.value());
         }
         else
         {
@@ -70,7 +70,7 @@ namespace cpu
 
         if(metadata.avg_data.is_initialized())
         {
-            avg_data = ToMatrix(metadata.avg_data.value());
+            m_avg_data = ToMatrix(metadata.avg_data.value());
         }
         else
         {
@@ -114,8 +114,8 @@ namespace cpu
             }
         }
 
-        avg_data = Eigen::MatrixXcd::Zero(ms.GetNumBaselines(), ms.GetNumPols());
-        BOOST_LOG_TRIVIAL(info) << "avg_data:" << avg_data.size() * sizeof(std::complex<double>) / (1024.0 * 1024.0 * 1024.0) << " GiB";
+        m_avg_data = Eigen::MatrixXcd::Zero(ms.GetNumBaselines(), ms.GetNumPols());
+        BOOST_LOG_TRIVIAL(info) << "avg_data:" << m_avg_data.size() * sizeof(std::complex<double>) / (1024.0 * 1024.0 * 1024.0) << " GiB";
 
         //select the first epoch only
         casacore::Vector<double> time = msmc->time().getColumn();
@@ -140,8 +140,8 @@ namespace cpu
 
         BOOST_LOG_TRIVIAL(info) << "Calculating PhaseMatrix A1";
         std::tie(m_A1, m_I1) = icrar::cpu::PhaseMatrixFunction(ToVector(a1), ToVector(a2), 0);
-        BOOST_LOG_TRIVIAL(trace) << pretty_matrix(m_A1);
 #ifdef TRACE
+        BOOST_LOG_TRIVIAL(trace) << pretty_matrix(m_A1);
         {
             std::ofstream file;
             file.open("A1.txt");
@@ -152,8 +152,8 @@ namespace cpu
 
         BOOST_LOG_TRIVIAL(info) << "Calculating PhaseMatrix A";
         std::tie(m_A, m_I) = icrar::cpu::PhaseMatrixFunction(ToVector(a1), ToVector(a2), -1);
-        BOOST_LOG_TRIVIAL(trace) << pretty_matrix(m_A);
 #ifdef TRACE
+        BOOST_LOG_TRIVIAL(trace) << pretty_matrix(m_A);
         {
             std::ofstream file;
             file.open("A.txt");
@@ -164,8 +164,8 @@ namespace cpu
 
         BOOST_LOG_TRIVIAL(info) << "Inverting PhaseMatrix A1";
         m_Ad1 = icrar::cpu::PseudoInverse(m_A1);
-        BOOST_LOG_TRIVIAL(trace) << pretty_matrix(m_Ad1);
 #ifdef TRACE
+        BOOST_LOG_TRIVIAL(trace) << pretty_matrix(m_Ad1);
         {
             std::ofstream file;
             file.open("Ad1.txt");
@@ -176,8 +176,8 @@ namespace cpu
 
         BOOST_LOG_TRIVIAL(info) << "Inverting PhaseMatrix A";
         m_Ad = icrar::cpu::PseudoInverse(m_A);
-        BOOST_LOG_TRIVIAL(trace) << pretty_matrix(m_Ad);
 #ifdef TRACE
+        BOOST_LOG_TRIVIAL(trace) << pretty_matrix(m_Ad);
         {
             std::ofstream file;
             file.open("Ad.txt");
@@ -220,24 +220,24 @@ namespace cpu
 
     void MetaData::SetDD(const icrar::MVDirection& direction)
     {
-        this->direction = direction;
+        this->m_direction = direction;
 
         Eigen::Vector2d polar_direction = icrar::ToPolar(direction); 
         m_constants.dlm_ra = polar_direction(0) - m_constants.phase_centre_ra_rad;
         m_constants.dlm_dec = polar_direction(1) - m_constants.phase_centre_dec_rad;
 
-        dd = Eigen::Matrix3d();
-        dd(0,0) = std::cos(m_constants.dlm_ra) * std::cos(m_constants.dlm_dec);
-        dd(0,1) = -std::sin(m_constants.dlm_ra);
-        dd(0,2) = std::cos(m_constants.dlm_ra) * std::sin(m_constants.dlm_dec);
+        m_dd = Eigen::Matrix3d();
+        m_dd(0,0) = std::cos(m_constants.dlm_ra) * std::cos(m_constants.dlm_dec);
+        m_dd(0,1) = -std::sin(m_constants.dlm_ra);
+        m_dd(0,2) = std::cos(m_constants.dlm_ra) * std::sin(m_constants.dlm_dec);
         
-        dd(1,0) = std::sin(m_constants.dlm_ra) * std::cos(m_constants.dlm_dec);
-        dd(1,1) = std::cos(m_constants.dlm_ra);
-        dd(1,2) = std::sin(m_constants.dlm_ra) * std::sin(m_constants.dlm_dec);
+        m_dd(1,0) = std::sin(m_constants.dlm_ra) * std::cos(m_constants.dlm_dec);
+        m_dd(1,1) = std::cos(m_constants.dlm_ra);
+        m_dd(1,2) = std::sin(m_constants.dlm_ra) * std::sin(m_constants.dlm_dec);
 
-        dd(2,0) = -std::sin(m_constants.dlm_dec);
-        dd(2,1) = 0;
-        dd(2,2) = std::cos(m_constants.dlm_dec);
+        m_dd(2,0) = -std::sin(m_constants.dlm_dec);
+        m_dd(2,1) = 0;
+        m_dd(2,2) = std::cos(m_constants.dlm_dec);
     }
 
     void MetaData::SetOldUVW(const std::vector<icrar::MVuvw>& uvw)
@@ -252,7 +252,7 @@ namespace cpu
         m_UVW.reserve(m_oldUVW.size());
         for(size_t n = 0; n < size; n++)
         {
-            m_UVW.push_back(m_oldUVW[n] * dd);
+            m_UVW.push_back(m_oldUVW[n] * m_dd);
         }
     }
 
@@ -267,8 +267,8 @@ namespace cpu
         && m_A1 == rhs.m_A1
         && m_I1 == rhs.m_I1
         && m_Ad1 == rhs.m_Ad1
-        && dd == rhs.dd
-        && avg_data == rhs.avg_data;
+        && m_dd == rhs.m_dd
+        && m_avg_data == rhs.m_avg_data;
     }
 
     bool Constants::operator==(const Constants& rhs) const