--- a/src/icrar/leap-accelerate/model/cuda/DeviceMetaData.cu
+++ b/src/icrar/leap-accelerate/model/cuda/DeviceMetaData.cu
@@ -34,9 +34,9 @@ namespace cuda
     : constants(metadata.GetConstants())
     , UVW(metadata.m_UVW)
     , oldUVW(metadata.m_oldUVW)
-    , dd(metadata.dd)
-    , direction(metadata.direction)
-    , avg_data(metadata.avg_data)
+    , dd(metadata.GetDD())
+    , direction(metadata.GetDirection())
+    , avg_data(metadata.m_avg_data)
     , A(metadata.GetA())
     , I(metadata.GetI())
     , Ad(metadata.GetAd())
@@ -64,9 +64,9 @@ namespace cuda
 
         oldUVW.ToHost(metadata.m_oldUVW);
         UVW.ToHost(metadata.m_UVW);
-        metadata.direction = direction;
-        metadata.dd = dd;
-        avg_data.ToHost(metadata.avg_data);
+        metadata.m_direction = direction;
+        metadata.m_dd = dd;
+        avg_data.ToHost(metadata.m_avg_data);
     }
 
     void DeviceMetaData::AvgDataToHost(Eigen::MatrixXcd& host) const