--- a/src/icrar/leap-accelerate/model/cuda/DeviceMetaData.h
+++ b/src/icrar/leap-accelerate/model/cuda/DeviceMetaData.h
@@ -108,6 +108,13 @@ namespace cuda
         void ToHost(icrar::cpu::MetaData& host) const;
         icrar::cpu::MetaData ToHost() const;
         void ToHostAsync(icrar::cpu::MetaData& host) const;
+
+        /**
+         * @brief Copies average data to host memory
+         * 
+         * @param host 
+         */
+        void AvgDataToHost(Eigen::MatrixXcd& host) const;
     };
 }
 }