Here done some cleanup around object encapsulation and a small improvement to cuda performance by only copying the avg_data buffer from gpu to cpu memory per direction. Benchmarking on my machine this improved:
* single direction SKA wall time 51.41s -> 50.53s
* multi direction SKA wall time 65.45s -> 57.68s