@rtobar 2020-11-02
I tested the changes now on hyades and indeed the phase rotation time goes down as expected :+1:. I left more comments on YAN-516.