@calgray 2020-11-03
I might change one of the CI builds to Debug configuration as I often forget debug compiles extra lines of source