@rtobar 2020-11-03
Yeap, good idea, but remember that pure debug builds fail because they take too long to run the test; so maybe just add the `-DNDEBUG` into the `CMAKE_CXX_FLAGS` in the travis `cmake` invocation.

Anyway, I tried the changes in my computer and now it builds, so I've just approved this.